requirements/dev.txt
-r base.txt

pytest-django==3.9.0
mock==4.0.2
model-bakery==1.1.0
factory-boy==2.12.0
